refactor(auth): consolidate auth modules into backend/services/auth/ package (audit A9, closes #154)
Audit cycle 2026-05-13 PR A9.

## Why

Four files in backend/services/ all named auth_* with overlapping
concerns:

  auth_manager.py     1794 LOC  AuthManager policy/state engine
  auth_service.py      272 LOC  AuthService request-time facade
  auth_services.py     680 LOC  4 sub-services (token/session/MFA/lockout)
  auth_repository.py  1087 LOC  Legacy AuthRepository (DB persistence)

Naming-by-counting-s is cognitive load tax for every reader, and the
runtime call chain compounds it -- main.py constructs AuthService,
registers it under registry key 'auth_manager', and every router calls
get_auth_manager().get_auth_manager() to unwrap the underlying engine.

## What changed

All four files moved under a single backend/services/auth/ package.
auth_services.py is split into per-class modules as the audit prompt
specifies:

  backend/services/auth/
    __init__.py        -- public re-exports, package-level docstring
                          documenting the layered AuthManager/AuthService
                          design (kept intentionally per ADR-0007)
    manager.py         -- AuthManager           (git mv from auth_manager.py)
    service.py         -- AuthService           (git mv from auth_service.py)
    repository.py      -- AuthRepository        (git mv from auth_repository.py)
    tokens.py          -- TokenService          (extracted from auth_services.py)
    sessions.py        -- SessionService        (extracted from auth_services.py)
    mfa.py             -- MfaService            (extracted from auth_services.py)
    lockout.py         -- LockoutService        (extracted from auth_services.py)

git mv was used for the 3 moves so blame is preserved. The 4 splits
are verbatim copies of the original class body; only the surrounding
imports and module docstring are new.

External callers updated (14 import sites across backend/ + tests/):
  backend.services.auth_manager    -> backend.services.auth.manager
  backend.services.auth_service    -> backend.services.auth.service
  backend.services.auth_services   -> backend.services.auth.{tokens,sessions,mfa,lockout}
  backend.services.auth_repository -> backend.services.auth.repository

## Explicitly out of scope (deferred follow-ups)

The audit prompt also calls for renaming the registry key 'auth_manager'
and eliminating the get_auth_manager().get_auth_manager() charade.
That work touches every auth router + every middleware that pulls auth
from DI and is intentionally deferred -- doing it in the same PR as the
file relocation would make this diff unreviewable. The follow-ups are
called out in ADR-0007 and the package __init__.py.

AuthManager and AuthService are NOT merged -- per the audit prompt's
stop-and-ask, they encapsulate genuinely different responsibilities
(policy engine vs request-time facade) and merging them would damage
testability. Documented in ADR-0007.

## Verification

  poetry run python -c 'from backend.services.auth import AuthManager, AuthService, ...'  -> OK
  poetry run python -c 'from backend.main import create_app; create_app()'                 -> OK (315 routes)
  poetry run pytest tests/test_auth_decorators tests/test_authentication \
                    tests/safety_validation_test tests/integrations/auth/* \
                    tests/test_pin_security tests/services/test_pin_manager_db \
                    tests/api/test_safety_pin_endpoints --no-cov                            -> 92 passed, 2 pre-existing failures (verified on origin/main)
  poetry run pyright backend                                                                -> 1443 errors (unchanged baseline)

Refs #145.
Closes #154.
