+++ backend/services/user_invitation_service.py
@@ -13,7 +13,7 @@ from datetime import UTC, datetime, timedelta
 from pydantic import BaseModel, EmailStr
 
 from backend.core.config import get_settings
-from backend.services.auth_manager import AuthManager
+from backend.services.auth.manager import AuthManager
 from backend.services.notification_manager import NotificationManager
 
 logger = logging.getLogger(__name__)
